Updated Cogency, Syntax, and Clone Explanation
The description said to install "NPM/Node." In the link that's provided, the author says that Node is being installed and specifically states that NPM is not being installed. It's more accurate to just say Node is being installed for now to avoid permission issues. Added colon for a list. Removed first person to create a consistent voice. Explained what cloning was because that's not covered up to this point in Foundations.  Made the capitalization consistent with the rest of the course.