Update Fibonacci solution, test to focus on arrays
Previous Fibonacci solution didn't use arrays, and since this is an
array-heavy section, it seemed better to have that be the method here.

Also, it accounts for entering 0 as an argument without having to add
any extra code, which takes care of some currently open issues.
Issues #192 and #236 are about the same thing.

I added a test for that as well.